Outlier in Rhandsontable

The outlier will highlighted in green and the word 'Outlier' will replace the value in the cell.

Usage

routlier_rh(data)

Arguments

data

filepath to data

Value

Return's an outlier dataset from the original dataset in an rhandonstable.

Examples

  routlier_rh(data = iris)
#> You have  1  outliers in your dataset