routlier_tukey

Routlier: Finds the Outliers in a data set utilizing the Tukey's Method:

routlier_tukey(data,type,outlier_type)

Arguments

data
filepath to data

type
select type of quatile method to use. Choices from 1 to 7.

outlier_type
select type of outlier selecting either "M" for mild or "E" for extreme.

Value

Return's an outlier dataset from the original dataset in a formattable table. The data returned is currently the numeric data only from the dataset.

Details

The word 'Outlier' will replace the value that is an outlier in the respective cell.

Examples

## Load the routlier library
   library(routlier)

## Look at outliers utilizng the Tukey method
  routlier_tukey(data = detroit,type = 7,outlier_type="M")

#> [1] "The IQR for column 1 is from: 448.875 : 162.355 and the overall IQR range is: 286.52"
#> [1] "The IQR for column 2 is from: 11.9 : -0.9 and the overall IQR range is: 12.8"
#> [1] "The IQR for column 3 is from: 688.55 : 444.15 and the overall IQR range is: 244.4"
#> [1] "The IQR for column 4 is from: 1654.1 : -637.1 and the overall IQR range is: 2291.2"
#> [1] "The IQR for column 5 is from: 1528.57 : -546.47 and the overall IQR range is: 2075.04"
#> [1] "The IQR for column 6 is from: 116.65 : 48.25 and the overall IQR range is: 68.4"
#> [1] "The IQR for column 7 is from: 648865.5 : 253109.5 and the overall IQR range is: 395756"
#> [1] "The IQR for column 8 is from: 1001.75 : 344.55 and the overall IQR range is: 657.2"
#> [1] "The IQR for column 9 is from: 334.05 : 40.05 and the overall IQR range is: 294"
#> [1] "The IQR for column 10 is from: 6.33 : 1.37 and the overall IQR range is: 4.96"
#> [1] "The IQR for column 11 is from: 234.33 : 86.09 and the overall IQR range is: 148.24"
#> [1] "The IQR for column 12 is from: 80.125 : -33.835 and the overall IQR range is: 113.96"
#> [1] "The IQR for column 13 is from: 60.295 : 34.495 and the overall IQR range is: 25.8"
#> [1] "The IQR for column 14 is from: 391.28 : 209.28 and the overall IQR range is: 182"
#> [1] "You have a total of 4 Outliers in your dataset"