Homicide in Detroit: The role of Firearms.

This is the data set called `DETROIT' in the book `Subset selection in regression' by Alan J. Miller published in the Chapman & Hall series of monographs on Statistics & Applied Probability, no. 40. The data are unusual in that a subset of three predictors can be found which gives a very much better fit to the data than the subsets found from the Efroymson stepwise algorithm, or from forward selection or backward elimination. The original data were given in appendix A of `Regression analysis and its application: A data-oriented approach' by Gunst & Mason, Statistics textbooks and monographs no. 24, Marcel Dekker. It has caused problems because some copies of the Gunst & Mason book do not contain all of the data, and because Miller does not say which variables he used as predictors and which is the dependent variable. (HOM was the dependent variable, and the predictors were FTP ... WE)

detroit

Format

A data frame with 13 rows and 14 variables:

FTP: Full-time police per 100,000 population

UEMP: UEMP - % unemployed in the population

MAN: MAN - number of manufacturing workers in thousands

LIC: LIC - Number of handgun licences per 100,000 population

GR: GR - Number of handgun registrations per 100,000 population

CLEAR: CLEAR - % homicides cleared by arrests

WM: WM - Number of white males in the population

NMAN: NMAN - Number of non-manufacturing workers in thousands

GOV: GOV - Number of government workers in thousands

HE: HE - Average hourly earnings

WE: WE - Average weekly earnings:

HOM: HOM - Number of homicides per 100,000 of population

ACC: ACC - Death rate in accidents per 100,000 population

ASR: ASR - Number of assaults per 100,000 population

Source

http://lib.stat.cmu.edu/datasets/detroit